Split CMake Linux builds (#400)
Simplifying the most commonly used build system entrypoint by isolating
all the Steam runtime static linking nonsense into its own dark corner
in the cellar.

Also paving the road to tackling Windows and macOS with cmake.

This changes the default command to prep the build, but that is
documented in the readme. Local experimentation most welcome during a
review. You might have to remove the untracked directory `CMakeFiles/`
and the untracked file`CMakeCache.txt` as the location of the main build
definition has changed and then `cmake buildsystem/linux-dynamic` should
generate you a `Makefile` for a debug build.